- Rendered more hooks than during the previous render - Restored styling?
