+++ src/components/ProgressiveDataUnlocks.ts
@@ -3,21 +3,21 @@
 // Logic for unlocking progressive data layers based on game state
 // Adjust thresholds and logic as needed for game balance and pacing.
 
-import { DataLayer } from './ProgressiveDataLayers';
+import { DataLayer } from "./ProgressiveDataLayers";
 
 // Example GameState type (adjust as needed)
 export interface GameState {
-    fieldsPhasedOut: number;
-    score: number;
-    year: number;
-    achievements: string[];
+  fieldsPhasedOut: number;
+  score: number;
+  year: number;
+  achievements: string[];
 }
 
 // Thresholds for unlocking layers
 const INTERMEDIATE_FIELDS = 4; // Number of fields to phase out for intermediate
-const ADVANCED_SCORE = 2000;   // Minimum score for advanced
-const ADVANCED_YEAR = 2035;    // Minimum year for advanced
-const EXPERT_ACHIEVEMENTS = ['expert-achievement', 'legendary-achievement']; // Both required for expert
+const ADVANCED_SCORE = 2000; // Minimum score for advanced
+const ADVANCED_YEAR = 2035; // Minimum year for advanced
+const EXPERT_ACHIEVEMENTS = ["expert-achievement", "legendary-achievement"]; // Both required for expert
 
 /**
  * Returns an array of unlocked data layers based on the current game state.
@@ -31,25 +31,22 @@ const EXPERT_ACHIEVEMENTS = ['expert-achievement', 'legendary-achievement']; // 
  * @returns Array of unlocked layer names in order
  */
 export function getUnlockedLayers(gameState: GameState): DataLayer[] {
-    const unlocked: DataLayer[] = ['basic'];
+  const unlocked: DataLayer[] = ["basic"];
 
-    // Intermediate: Unlock after 4 fields phased out
-    if (gameState.fieldsPhasedOut >= INTERMEDIATE_FIELDS) {
-        unlocked.push('intermediate');
-    }
+  // Intermediate: Unlock after 4 fields phased out
+  if (gameState.fieldsPhasedOut >= INTERMEDIATE_FIELDS) {
+    unlocked.push("intermediate");
+  }
 
-    // Advanced: Unlock after BOTH score and year thresholds
-    if (
-        gameState.score >= ADVANCED_SCORE &&
-        gameState.year >= ADVANCED_YEAR
-    ) {
-        unlocked.push('advanced');
-    }
+  // Advanced: Unlock after BOTH score and year thresholds
+  if (gameState.score >= ADVANCED_SCORE && gameState.year >= ADVANCED_YEAR) {
+    unlocked.push("advanced");
+  }
 
-    // Expert: Unlock after BOTH required achievements
-    if (EXPERT_ACHIEVEMENTS.every(a => gameState.achievements.includes(a))) {
-        unlocked.push('expert');
-    }
+  // Expert: Unlock after BOTH required achievements
+  if (EXPERT_ACHIEVEMENTS.every((a) => gameState.achievements.includes(a))) {
+    unlocked.push("expert");
+  }
 
-    return unlocked;
-} 
+  return unlocked;
+}
