+++ src/routes/__root.tsx
@@ -3,6 +3,7 @@ import {
   Link,
   Outlet,
   useRouterState,
+  useNavigate,
 } from "@tanstack/react-router";
 import React, { useState, useEffect } from "react";
 import "./RootLayout.css";
@@ -11,134 +12,9 @@ import {
   useGameStats,
   useGameState,
 } from "../context/GameStateContext";
-
-// Mobile Navigation Component
-const MobileNavigation: React.FC<{ currentPath: string }> = ({
-  currentPath,
-}) => {
-  // Use try-catch to handle context not being available
-  let gameStats;
-  let dispatch;
-  try {
-    gameStats = useGameStats();
-    const { dispatch: gameDispatch } = useGameState();
-    dispatch = gameDispatch;
-  } catch {
-    // Fallback values when context is not available
-    gameStats = {
-      totalBudgetSpent: 0,
-      fieldsRemaining: 50,
-      selectedFieldsCount: 0,
-      totalEmissionsReduced: 0,
-    };
-    dispatch = () => {};
-  }
-
-  const navItems = [
-    {
-      id: "game",
-      icon: "🎮",
-      label: "Spill",
-      path: "/phase-out-game",
-      active: currentPath === "/phase-out-game",
-    },
-    {
-      id: "investments",
-      icon: "💰",
-      label: "Investering",
-      path: "/investments",
-      active: currentPath === "/investments",
-    },
-    {
-      id: "stats",
-      icon: "📊",
-      label: "Statistikk",
-      path: "/stats",
-      active: currentPath === "/stats",
-    },
-    // {
-    //   id: "table",
-    //   icon: "📋",
-    //   label: "Data",
-    //   path: "/productiontable",
-    //   active: currentPath === "/productiontable",
-    // },
-
-    {
-      id: "about",
-      icon: "ℹ️",
-      label: "Om",
-      path: "/about",
-      active: currentPath === "/about",
-    },
-  ];
-
-  const handleRestart = () => {
-    if (
-      window.confirm(
-        "Er du sikker på at du vil starte spillet på nytt? All fremgang vil gå tapt.",
-      )
-    ) {
-      localStorage.removeItem("phaseOutVillage_gameState");
-      dispatch({ type: "RESTART_GAME" });
-    }
-  };
-
-  return (
-    <>
-      {/* Game Status Bar - Always visible on mobile */}
-      <div className="mobile-status-bar">
-        <div className="status-item">
-          <span className="status-icon">💰</span>
-          <span className="status-value">
-            {((100000000 - gameStats.totalBudgetSpent) / 1000000).toFixed(1)}M
-            kr
-          </span>
-        </div>
-        <div className="status-item">
-          <span className="status-icon">🛢️</span>
-          <span className="status-value">{gameStats.fieldsRemaining}</span>
-        </div>
-        <div className="status-item">
-          <span className="status-icon">🌱</span>
-          <span className="status-value">{gameStats.selectedFieldsCount}</span>
-        </div>
-        <div className="status-item">
-          <span className="status-icon">💨</span>
-          <span className="status-value">
-            {gameStats.totalEmissionsReduced.toFixed(1)}Mt
-          </span>
-        </div>
-      </div>
-
-      {/* Bottom Navigation */}
-      <nav className="mobile-bottom-nav">
-        {navItems.map((item) => (
-          <Link
-            key={item.id}
-            to={item.path}
-            className={`nav-item ${item.active ? "active" : ""}`}
-          >
-            <div className="nav-icon-container">
-              <span className="nav-icon">{item.icon}</span>
-            </div>
-            <span className="nav-label">{item.label}</span>
-          </Link>
-        ))}
-        {/* Restart button */}
-        <div className="mobile-restart-container">
-          <button
-            onClick={handleRestart}
-            className="mobile-restart-button"
-            title="Start på nytt"
-          >
-            🔄
-          </button>
-        </div>
-      </nav>
-    </>
-  );
-};
+import BottomNavBar from "../components/Navigation/BottomNavBar";
+import MobileStatusBar from "../components/Navigation/MobileStatusBar";
+import DesktopStatusBar from "../components/Navigation/DesktopStatusBar";
 
 // Desktop Navigation Component
 const DesktopNavigation: React.FC<{ currentPath: string }> = ({
@@ -158,7 +34,6 @@ const DesktopNavigation: React.FC<{ currentPath: string }> = ({
     { label: "Spill", path: "/phase-out-game" },
     { label: "Investeringer", path: "/investments" },
     { label: "Statistikk", path: "/stats" },
-    // { label: "Produksjonstabell", path: "/productiontable" },
     { label: "Om", path: "/about" },
   ];
 
@@ -186,7 +61,8 @@ const DesktopNavigation: React.FC<{ currentPath: string }> = ({
             <Link
               key={item.path}
               to={item.path}
-              className={`nav-link ${currentPath === item.path ? "active" : ""}`}
+              className={`nav-link${currentPath === item.path ? " active" : ""}`}
+              // No onClick or preventDefault, let router handle navigation
             >
               {item.label}
             </Link>
@@ -209,6 +85,8 @@ const RootLayout: React.FC = () => {
   const router = useRouterState();
   const currentPath = router.location.pathname;
   const [isMobile, setIsMobile] = useState(false);
+  const { dispatch } = useGameState();
+  const navigate = useNavigate();
 
   // Detect mobile/desktop
   useEffect(() => {
@@ -221,18 +99,73 @@ const RootLayout: React.FC = () => {
     return () => window.removeEventListener("resize", checkIsMobile);
   }, []);
 
+  // BottomNavBar items for mobile
+  const mobileNavItems = [
+    {
+      id: "game",
+      icon: "🎮",
+      label: "Spill",
+      active: currentPath === "/phase-out-game",
+      onClick: () => navigate({ to: "/phase-out-game" }),
+    },
+    {
+      id: "investments",
+      icon: "💰",
+      label: "Investering",
+      active: currentPath === "/investments",
+      onClick: () => navigate({ to: "/investments" }),
+    },
+    {
+      id: "stats",
+      icon: "📊",
+      label: "Statistikk",
+      active: currentPath === "/stats",
+      onClick: () => navigate({ to: "/stats" }),
+    },
+    {
+      id: "about",
+      icon: "ℹ️",
+      label: "Om",
+      active: currentPath === "/about",
+      onClick: () => navigate({ to: "/about" }),
+    },
+    {
+      id: "restart",
+      icon: "🔄",
+      label: "Start på nytt",
+      active: false,
+      onClick: () => {
+        if (
+          window.confirm(
+            "Er du sikker på at du vil starte spillet på nytt? All fremgang vil gå tapt."
+          )
+        ) {
+          localStorage.removeItem("phaseOutVillage_gameState");
+          dispatch({ type: "RESTART_GAME" });
+        }
+      },
+      ariaLabel: "Start på nytt",
+    },
+  ];
+
   return (
     <div className="root-layout">
       {/* Desktop Navigation */}
       {!isMobile && <DesktopNavigation currentPath={currentPath} />}
+
+      {/* Desktop Status Bar */}
+      {!isMobile && <DesktopStatusBar />}
 
       {/* Main Content */}
       <main className={`main-content ${isMobile ? "mobile" : "desktop"}`}>
         <Outlet />
       </main>
 
+      {/* Mobile Status Bar */}
+      {isMobile && <MobileStatusBar />}
+
       {/* Mobile Navigation */}
-      {isMobile && <MobileNavigation currentPath={currentPath} />}
+      {isMobile && <BottomNavBar items={mobileNavItems} />}
     </div>
   );
 };
