+++ src/routes/phase-out-game.tsx
@@ -1,12 +1,11 @@
 import React from "react";
 import { createFileRoute } from "@tanstack/react-router";
-import CompactMapPage from "../pages/CompactMapPage/CompactMapPage";
+import MapPage from "../pages/MapPage/MapPage";
 
 export const Route = createFileRoute("/phase-out-game")({
   component: RouteComponent,
 });
 
 function RouteComponent() {
-  // return <PhaseOutMapPage />;
-  return <CompactMapPage />;
+  return <MapPage />;
 }
